mysite/requirements.txt
pillow==10.4.0
Django==5.0.6
psycopg2-binary==2.9.9
djoser==2.2.2
djangorestframework==3.15.1
